requirements.txt
docopt
pyyaml
splinter
